requirements_full.txt
pandas>=2.0.0,<3.0.0
numpy>=1.24.0
faker>=19.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Apache Kafka
kafka-python>=2.0.2

# Apache Spark (PySpark)
pyspark>=3.4.0

# Apache Airflow
apache-airflow>=2.7.0
apache-airflow-providers-apache-spark>=4.1.0
apache-airflow-providers-postgres>=5.6.0

# Database drivers
psycopg2-binary>=2.9.7

# Additional utilities
requests>=2.31.0
